♻️ refactor : 프로젝트 폴더 구조 리팩토링
before
  components/MyPage 폴더에
  모든 스타일되지 않은 컴포넌트 파일을 위치시키고
  components/MyPage/MyPage.styles 폴더에
  모든 스타일된 컴포넌트 파일을 위치시킴

after
  components/MyPage 폴더에
  컴포넌트 별로 폴더를 만들고
  그 안에 기능을 구현한 컴포넌트와
  해당 기능 컴포넌트를 스타일링한 컴포넌트를 위치시킴

  ex) components/MyPage/PostList 폴더에
    스타일링 되지 않은 순수 기능 컴포넌트인
    PostList.jsx파일과
    PostList.jsx를 스타일링한 PostList.styled.jsx 파일을
    함께 위치시킴

+ devutil.js 파일을 src/mockdatas 폴더로 이동
