+++ src/components/MyPage/MyPage.styles/LeftSection.styled.jsx
@@ -1,13 +1,13 @@
 import styled from 'styled-components';
-import { gap, padding } from './common-values';
+import { gap, padding } from './style-common-values';
 
 const LeftSection = styled.section`
   width: 25%;
 
   display: flex;
   flex-direction: column;
-  gap: ${gap};
-  padding: ${padding};
+  gap: ${gap}px;
+  padding: ${padding}px;
 
   background-color: palegreen; //영역 구분용 컬러
 `;
