+++ src/components/MyPage/MyPage.styles/RightSection.styled.jsx
@@ -1,13 +1,13 @@
 import styled from 'styled-components';
-import { gap, padding } from './common-values';
+import { gap, padding } from './style-common-values';
 
 const RightSection = styled.section`
   width: 75%;
 
   display: flex;
   flex-direction: column;
-  gap: ${gap};
-  padding: ${padding};
+  gap: ${gap}px;
+  padding: ${padding}px;
 
   background-color: palevioletred; //영역 구분용 컬러
 `;
