--- src/components/MyPage/MyPage.styles/common-values.js
@@ -1,4 +1,0 @@
-const gap = '10px';
-const padding = '10px';
-
-export { gap, padding };
