+++ src/components/MyPage/MyPage.styles/style-common-values.js
@@ -1,0 +1,4 @@
+const gap = '10';
+const padding = '10';
+
+export { gap, padding };
