+++ src/components/MyPage/Summary/Summary.styled.jsx
@@ -1,15 +1,15 @@
 import styled from 'styled-components';
-import { gap, padding } from './common-values';
-
-import Summary from '../Summary';
+//import { gap, padding } from './../common-values';
+import Summary from './Summary';
+import { gap, padding } from '../MyPage.styles/style-common-values';
 
 const StSummary = styled(Summary)`
   height: 15%;
 
   display: flex;
   justify-content: space-around;
-  gap: ${gap};
-  padding: ${padding};
+  gap: ${gap}px;
+  padding: ${padding}px;
 
   background-color: var(--box-color);
