+++ src/pages/MyPage/MyPage.jsx
@@ -1,21 +1,23 @@
-//사용 컴포넌트
+//순수 스타일용 컴포넌트
 import Main from '../../components/MyPage/MyPage.styles/Main.styled';
 import LeftSection from '../../components/MyPage/MyPage.styles/LeftSection.styled';
 import RightSection from '../../components/MyPage/MyPage.styles/RightSection.styled';
-import StProfileImg from '../../components/MyPage/MyPage.styles/ProfileImg.styled';
-import StSummary from '../../components/MyPage/MyPage.styles/Summary.styled';
-import StPostList from '../../components/MyPage/MyPage.styles/PostList.styled';
 import StSide from '../../components/MyPage/MyPage.styles/Side.styled';
 import StLink from '../../components/MyPage/MyPage.styles/Link.styled';
 import Button from '../../components/MyPage/MyPage.styles/Button.styled';
+
+//스타일된 기능용 컴포넌트
+import StSummary from '../../components/MyPage/Summary/Summary.styled';
+import StProfileImg from '../../components/MyPage/ProfileImg/ProfileImg.styled';
+import StPostList from '../../components/MyPage/PostList/PostList.styled';
 
 //리액트 라이브러리
 import { useState } from 'react';
 
 //가짜 데이터
-import { loadMyPageMockData } from '../../components/MyPage/devutil';
+import { loadMockData } from '../../mockdatas/devutil';
 
-const { mockUserList, mockPostList, mockLoginedUser } = loadMyPageMockData();
+const { mockUserList, mockPostList, mockLoginedUser } = loadMockData();
 
 function MyPage() {
   const [userList, setUserList] = useState(mockUserList);
